fix: shutdown has to be the last request tested
When a Shutdown request is received, the code will set a flag to signal it
and all requests after it are going to fail and respond with a fail message.
Since the code runs on every handle invocation, processing a Shutdown first
makes it so all subsequent handles send a message to the client. Processing
the shutdown request last is the easiest solution to this problem.