feat: go to definition
The open file is now parsed and information about declarations and definitions
are extracted, allowing "go to definition" to work. It supports jumping inside
the document, in the path and inside packages. There is code in place to deal
with classes and class folders too, but I don't care about those so they are
mostly useless for now.