feat: references
Listing references to symbol now works for variables, functions and classes.
Again, the code for classes is complete in the sense that, if I ever correctly
parse a class and generate references for them, then listing such references
will work correctly.

There is no reference to namespaces/packages as that does not sound meaningful to
me. References to structure fields works, although how accurate it is depends on
how accurate the references were generated by the refdef.rs.

To make references work, a lot of changes touching the parsing of files and how
data is stored were made. Now every file should only be listed once in the global
structure, files get rescanned to fix cross-refs in the worker thread and the
workspace structures are kept up-to-date with changes being made in the open files.