fix: scopes and lambda defref
When refactoring I ended up changing some code in the wrong way, which messed up
the scopes (inverted it, which makes no difference without nested functions so
I didn't notice) and added some logic to assignments to avoid declaring a
reference to a definition in the same assignment. This fixes both so now all the
scopes are right and defref inside lambdas work as expected.