fix: comments can span multiple lines, so the lenght is the byte span
For the other tokens it is not a problem since they are not allowed to span multiple
lines, however comments can, so simply using the chars won't do it. I'm not sure if
this is the correct way of doing it, it should probably be calculated by getting
the node's text and counting the number of chars, but this is not a very important
token for semantics, so whatever. Tree-Sitter is getting it right in the client anyway.