feat: diagnostic for undefined variable usage
Alongside some changes on how the first appearance of a struct is interpreted,
since it has to be a definition so the diagnostic don't bother you about it.
Also changes the way the `end` keyword is handled inside function calls and ranges
so that they are not marked as undefined variables either.